Add finalize_purchase

+++ purchase/views.py
@@ -1,10 +1,11 @@
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.http import HttpResponse
 from django.shortcuts import render, redirect, reverse
 from django.views.generic.base import View
 
 from product.models import Product
-from .forms import EditPurchaseItemForm
+from .forms import EditPurchaseItemForm, FinalizePurchaseForm
 from .models import Purchase, PurchaseItem
 
 
@@ -18,7 +19,7 @@ def add_item_to_purchase(request, product_id):
         purchase = Purchase.objects.create(user=request.user, state='N')
 
     try:
-        product = Product.objects.get(id=product_id)
+        product = Product.objects.get(id=product_id, is_available=True)
     except Product.DoesNotExist:
         return HttpResponse(status=404)
 
@@ -60,13 +61,16 @@ def show_purchase(request):
     })
 
 
-class EditPurchaseItemView(View):
+class EditPurchaseItemView(LoginRequiredMixin, View):
     @staticmethod
     def get(request, purchase_item_id):
         try:
             purchase_item = PurchaseItem.objects.get(id=purchase_item_id)
         except PurchaseItem.DoesNotExist:
             return HttpResponse(status=404)
+
+        if purchase_item.purchase.user != request.user:
+            return HttpResponse(status=403)
 
         form = EditPurchaseItemForm(instance=purchase_item)
         return render(request, 'purchase/edit_purchase_item.html',
@@ -84,18 +88,64 @@ class EditPurchaseItemView(View):
         except PurchaseItem.DoesNotExist:
             return HttpResponse(status=404)
 
+        if purchase_item.purchase.user != request.user:
+            return HttpResponse(status=403)
+
         form = EditPurchaseItemForm(request.POST, instance=purchase_item)
         if form.is_valid():
             form.save()
             return redirect(reverse('purchase:show_purchase'))
 
+        return render(request, 'purchase/edit_purchase_item.html',
+                      {
+                          'form': form,
+                          'purchase_item': {
+                              'id': purchase_item.id,
+                              'name': purchase_item.product.name,
+                          }})
 
+
+@login_required
 def delete_purchase_item(request, purchase_item_id):
     try:
         purchase_item = PurchaseItem.objects.get(id=purchase_item_id)
     except PurchaseItem.DoesNotExist:
         return HttpResponse(status=404)
 
+    if purchase_item.purchase.user != request.user:
+        return HttpResponse(status=403)
+
     purchase_item.delete()
     return redirect(reverse('purchase:show_purchase'))
 
+
+class FinalizePurchaseView(LoginRequiredMixin, View):
+    @staticmethod
+    def get(request):
+        try:
+            purchase = Purchase.objects.get(user=request.user, state='N')
+        except Purchase.DoesNotExist:
+            return HttpResponse(status=404)
+
+        form = FinalizePurchaseForm(instance=purchase)
+        return render(request, 'purchase/finalize_purchase.html', {
+            'form': form,
+        })
+
+    @staticmethod
+    def post(request):
+        try:
+            purchase = Purchase.objects.get(user=request.user, state='N')
+        except Purchase.DoesNotExist:
+            return HttpResponse(status=404)
+
+        form = FinalizePurchaseForm(request.POST, instance=purchase)
+        if form.is_valid():
+            purchase = form.save(commit=False)
+            purchase.state = 'D'
+            purchase.save()
+            return redirect(reverse('product_list'))
+
+        return render(request, 'purchase/finalize_purchase.html', {
+            'form': form,
+        })
